Fix some display issues when using windows. To do this, it was creates a geturi_local method that display the URI in the local format (local means using \ in Windows machine).
Now both clients internally uses always the posix format. local.py translates to local (windows) path format when needs to go to the filesystem.

Fix to allow honoring the environment variable REQUESTS_CA_BUNDLE to search for the certification files for the SSL connection. This is the standard variable to use. For some reason BOTO3 doesn't honor it, but urllib3 do, so many other connections APIs.